add some low hanging test coverage
